# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "ALAHLIFC API",
        "version": "1.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "get all products",
                "operationId": "products",
                "parameters": [
                    {
                        "name": "num_page",
                        "in": "formData",
                        "description": " number of page start from zero ( 0 )",
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "formData",
                        "description": " limit is number products will send in each time default ( 12 )",
                        "type": "number"
                    },
                    {
                        "name": "val_sort",
                        "in": "formData",
                        "description": "default is sort by ID (price_asc  - price_desc - rate)",
                        "type": "string"
                    },
                    {
                        "name": "lang",
                        "in": "formData",
                        "description": "default ar (ar - en)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
...
}